Fix plenty of reference problems in completion stuff.
* codegen/m4/convert_gtksourceview.m4: By default conversions don't take
additional references - this normally is not needed, as it should be done
explicitly when wrapping vfuncs and signals. Also, this file needs
a cleanup - maybe some conversions won't need to sit there.
* gtksourceview/src/sourcebuffer.hg:
* gtksourceview/src/sourcecompletion.hg:
* gtksourceview/src/sourcecompletionproposal.hg:
* gtksourceview/src/sourcecompletionprovider.hg:
* gtksourceview/src/sourcegutter.hg: Added explicit conversion for signal
and vfunc wrappers, so they take an additional reference to avoid premature
object disposal due to unreffing it upon temporary wrapper destruction.